+++ src/Presenters/CliPresenter.php
@@ -5,19 +5,20 @@ namespace Thinktomorrow\Vine\Presenters;
 use RecursiveArrayIterator;
 use RecursiveTreeIterator;
 
-class CliPresenter extends ArrayPresenter implements Presenter
+class CliPresenter extends ArrayPresenter
 {
     public function render()
     {
         $output = PHP_EOL;
-        $result = parent::render();
 
-        $iterator = new RecursiveArrayIterator($result);
+        $collection = $this->collection->toArray();
+
+        $iterator = new RecursiveArrayIterator($collection);
         $treeIterator = (new RecursiveTreeIterator(
             $iterator,
             RecursiveTreeIterator::BYPASS_KEY,
-            \CachingIterator::CATCH_GET_CHILD,
-            RecursiveTreeIterator::SELF_FIRST
+            RecursiveTreeIterator::CATCH_GET_CHILD,
+            RecursiveTreeIterator::SELF_FIRST,
         ));
 
         foreach ($treeIterator as $key => $value) {