+++ tests/Fixtures/FixtureSource.php
@@ -6,7 +6,7 @@ use Thinktomorrow\Vine\DefaultNode;
 use Thinktomorrow\Vine\Node;
 use Thinktomorrow\Vine\Source;
 
-class FixtureSource implements Source
+class FixtureSource
 {
     private $flatten;
 
@@ -21,23 +21,23 @@ class FixtureSource implements Source
         $this->flatten = require __DIR__.'/'.$filename;
     }
 
-    public function nodeEntries(): array
+    public function get(): array
     {
         return $this->flatten;
     }
-
-    public function nodeKeyIdentifier(): string
-    {
-        return 0;
-    }
-
-    public function nodeParentKeyIdentifier(): string
-    {
-        return 1;
-    }
-
-    public function createNode($entry): Node
-    {
-        return new DefaultNode($entry);
-    }
+//
+//    public function nodeKeyIdentifier(): string
+//    {
+//        return 0;
+//    }
+//
+//    public function nodeParentKeyIdentifier(): string
+//    {
+//        return 1;
+//    }
+//
+//    public function createNode($entry): Node
+//    {
+//        return new DefaultNode($entry);
+//    }
 }