+++ tests/Presenters/CliPresenterTest.php
@@ -3,6 +3,7 @@
 namespace Thinktomorrow\Vine\Tests\Presenters;
 
 use PHPUnit\Framework\TestCase;
+use Thinktomorrow\Vine\NodeCollection;
 use Thinktomorrow\Vine\NodeCollectionFactory;
 use Thinktomorrow\Vine\Presenters\CliPresenter;
 use Thinktomorrow\Vine\Source;
@@ -13,19 +14,16 @@ class CliPresenterTest extends TestCase
     /** @test */
     public function it_can_represent_tree_in_terminal()
     {
-        $tree = (new NodeCollectionFactory())->fromSource($this->getTranslation());
+        $tree = NodeCollection::fromIterable($this->getTranslation());
 
         $output = (new CliPresenter())->collection($tree)->render();
-
+die(var_dump($output));
         $this->assertIsString($output);
         $this->assertStringStartsWith('|-root-1', trim($output, PHP_EOL));
     }
 
-    /**
-     * @return Source
-     */
-    private function getTranslation(): Source
+    private function getTranslation(): iterable
     {
-        return new FixtureSource('default');
+        return (new FixtureSource('default'))->get();
     }
 }